Issue #19 – work in progress * using standardized-audio-context package to improve browser compatibility * audio context initialized after user interaction to prevent blocking by some browsers
