+++ packages/backend/src/socket/index.js
@@ -3,6 +3,7 @@ const ss = require("socket.io-stream");
 const { Input } = require("osx-audio");
 
 const {
+  PLAY,
   STATS_NEW_CLAP,
   STATS_NEW_SHARED_CLAPS,
   STATS_NEW_BOO,
@@ -19,9 +20,12 @@ module.exports = {
     io.set("origins", "*:*");
 
     io.on("connection", socket => {
-      const stream = ss.createStream();
-      ss(socket).emit("audio-stream", stream);
-      audioInput.pipe(stream);
+      socket.on(PLAY, () => {
+        console.log("USER CLICKED PLAY!");
+        const stream = ss.createStream();
+        ss(socket).emit("audio-stream", stream);
+        audioInput.pipe(stream);
+      });
       socket.on(STATS_NEW_CLAP, () => {
         console.log("CLAPPED!");
         loop.stats.claps++;
