+++ packages/frontend/package.json
@@ -10,7 +10,8 @@
     "react-dom": "^16.8.5",
     "react-scripts": "3.1.0",
     "socket.io-client": "^2.2.0",
-    "socket.io-stream": "^0.9.1"
+    "socket.io-stream": "^0.9.1",
+    "standardized-audio-context": "^21.2.1"
   },
   "scripts": {
     "start": "react-scripts start",
