+++ packages/frontend/src/App.js
@@ -1,10 +1,9 @@
 import React, { Component } from "react";
 import io from "socket.io-client";
-import ss from "socket.io-stream";
 import Lamp from "./Lamp";
 import ClapButton from "./ClapButton";
+import PlayButton from "./PlayButton";
 import Emoji from "./Emoji";
-import { getAudioContext, withWaveHeader } from "./utils";
 import styles from "./App.module.css";
 import BooButton from "./BooButton";
 
@@ -12,11 +11,6 @@ const { protocol, hostname } = window.location;
 const serverUrl = `${protocol}//${hostname}:3001`;
 const socket = io(serverUrl);
 
-/* delay to avoid timing issues */
-const delay = 1;
-const samplingRate = 44100;
-const chunkSize = 1024;
-const chunkDuration = chunkSize / samplingRate;
 
 class App extends Component {
   state = { claps: 0, boos: 0 };
@@ -28,24 +22,6 @@ class App extends Component {
   handleBoos = boos => {
     this.setState({ boos });
   };
-
-  componentDidMount() {
-    ss(socket).on("audio-stream", stream => {
-      const audioContext = getAudioContext().audioContext;
-      let nextTime = delay;
-      console.log("receiving audio stream");
-      stream.on("data", async data => {
-        const audioBufferChunk = await audioContext.decodeAudioData(
-          withWaveHeader(data)
-        );
-        const source = audioContext.createBufferSource();
-        source.buffer = audioBufferChunk;
-        source.connect(audioContext.destination);
-        source.start(nextTime);
-        nextTime += chunkDuration;
-      });
-    });
-  }
 
   render() {
     const { claps, boos } = this.state;
@@ -70,6 +46,7 @@ class App extends Component {
         <div className={styles.actions}>
           <ClapButton socket={socket} onClaps={this.handleClaps} />
           <BooButton socket={socket} onBoos={this.handleBoos} />
+          <PlayButton socket={socket} />
         </div>
       </div>
     );
