--- packages/frontend/src/utils/getAudioContext.js
@@ -1,6 +1,0 @@
-export default () => {
-  const MyAudioContext = window.AudioContext || window.webkitAudioContext;
-  const audioContext = new MyAudioContext();
-  const analyser = audioContext.createAnalyser();
-  return { audioContext, analyser };
-};
