+++ packages/frontend/src/utils/index.js
@@ -1,7 +1,5 @@
-import getAudioContext from "./getAudioContext";
 import withWaveHeader from "./withWaveHeader";
 
 export {
-  getAudioContext,
   withWaveHeader
 };
